Fixed bug #2224.
The problem was in the keystore, which did not contain the certificate
used in https://bitmask.net/ca.cert.

For the future, we should be able to easily manage this kind of
certificate upgrade/renewal without having to search again
(http://blog.crazybob.org/2010/02/android-trusting-ssl-certificates.html)
for the keytool command to import a certificate. Script?