Removed unused methods and variables.
Variables from LeapSRPSession were there because I used it while testing
srp calculations, comparing that strings with the ones from javascript.

Unused method from ProviderAPI was there because I foresee I'll have to
implement it in the future, but I've removed it since it's already in
the history.

This fixes #2781.