ConfigurationWizard backstack and exit behavior.
Feature #3023 first solution.

I don't like to tell anything to the user when s/he wants to exit
pressing the back button. I assume s/he is saying: "Wanna go out!!"

I've implemented the behaviour written in UI Rev II.

If a user selects a new provider when coming back from Switch Provider,
the previous provider is forgotten and we assume s/he wants to change of
provider but he is not sure to which. That means that previous provider
configuration will be removed (that means there is no provider selected
and Dashboard will not be launched before a new provider is selected).