If IOException, use current provider CA cert.
If the CA cert was correctly downloaded, we assume it can be used to
validate the string download.

If CA cert cannot validate that connection, then if the trust completely
checkbox was checked it will try with the other methods. If it was not
checked, a certificate error is shown (telling the user the
provider is not trusted).