Added android:logo, using the same as android:icon
For the moment, I don't see any problem on using the icon file we
have as the app's logo, although it's already used

Android Action Bar UI guide says that the logo should be wider and not
contain any text, while icon could. Since our icon is wide enough (from
my point of view) and does not contain any text, I think it's
appropriate to use it as a logo too.