Simulate network down.
The tests say they failed, but it's because the adb server is no longer
able to communicate with the emulator due to switching network
connectivity off (and being the adb server connected to a local tcp port
of the emulator, this port dissapears once connectivity is down).

To see if all tests passed, take a look at the logcat.
Once tests are finished, issue "adb kill-server" from the development
machine, and then "adb logcat".