An empty certificate is valid.
If there is no certificate, there is no need to fix it. Right now, we
use this class just to know if we need to update the certificate or
redownload it. In case the certificate is corrupted, then it won't be
empty but will contain an error message. If there is no
certificate (e.g., because the user hasn't logged in and the provider
doesn't provide anon vpn), then the "certificate" is ok but an upper
layer of the code will have to handle the situation (in this case, eip
will detect that there is no certificate and that the user needs to log
in, prompting it to do so).