Rename user session fragment, add icon resource.
The FabButton doesn't scale the icon as big as I want it to be, and it
doesn't let me set the size to "wrap_content" because the library sets
the sizes to match_parent.
I'm going to try to modify my fork of FabButton to see if I can obtain
what I want in its demo: just an icon and the progress indicator, the
icon as big as possible.